Move complexity class to statistics
move documentation to more appropriate locations
add deprecation warning to complexity_pdf
start implementing time_histogram for spread > 0
and epoch for spread = 0
